Changed 'CashRegister' to 'TaxCalculator' to better define it's function and created new CashRegister class which will contain, among other things, a TaxCalculator instance
