+++ TaxCalculator/src/TaxCalculator/Products/Product.java
@@ -4,19 +4,19 @@ import java.math.BigDecimal;
 
 public abstract class Product {
 
-    private final BigDecimal price;
-    private final String description;
+    private final BigDecimal PRICE;
+    private final String DESCRIPTION;
 
-    public Product(double price, String description) {
-        this.price = new BigDecimal(String.valueOf(price));
-        this.description = description;
+    public Product(double PRICE, String DESCRIPTION) {
+        this.PRICE = new BigDecimal(String.valueOf(PRICE));
+        this.DESCRIPTION = DESCRIPTION;
     }
 
     public BigDecimal getPrice() {
-        return this.price;
+        return this.PRICE;
     }
 
     public String getDescription() {
-        return this.description;
+        return this.DESCRIPTION;
     }
 }
