+++ TaxCalculator/src/TaxCalculator/TaxCalculator.java
@@ -1,0 +1,43 @@
+package TaxCalculator;
+
+import java.math.BigDecimal;
+import java.math.MathContext;
+
+public class TaxCalculator {
+
+    private TaxCalculator() {
+
+    }
+
+    public static TaxCalculator getInstance() {
+        return new TaxCalculator();
+    }
+
+    public BigDecimal calculateTaxRate(boolean isTaxExempt, boolean isImported) {
+        int nonExemptRate = 10;
+        int importedRate = 5;
+        double taxRate = 0;
+        if (!isTaxExempt) {
+            taxRate += nonExemptRate;
+        }
+        if (isImported) {
+            taxRate += importedRate;
+        }
+        return new BigDecimal(String.valueOf(taxRate/100));
+    }
+
+    public BigDecimal calculateTax(BigDecimal price, BigDecimal taxRate) {
+        MathContext mc = new MathContext(2);
+        BigDecimal tax = new BigDecimal(String.valueOf(price.multiply(taxRate)));
+        tax = tax.multiply(new BigDecimal(String.valueOf(20)));
+        double newTax = Math.ceil(tax.doubleValue());
+        tax = new BigDecimal(String.valueOf(newTax)).divide(new BigDecimal(String.valueOf(20)));
+        tax.round(mc);
+        return tax;
+    }
+
+    public BigDecimal calculateTotalPrice(BigDecimal price, BigDecimal tax) {
+        return price.add(tax);
+    }
+}
+
